Зробив ще трішки декомпозиції віджетів
