+++ packages/mobile_app/lib/common/card/big_card/big_card.dart
@@ -1,0 +1,95 @@
+import 'package:flutter/material.dart';
+import 'package:my_horoskope/app_global.dart';
+import 'package:my_horoskope/common/card/big_card/type/prophecy.dart';
+import 'package:my_horoskope/common/card/big_card/type/tarrot.dart';
+import 'package:my_horoskope/logic/cards/cards_logic.dart';
+import 'package:symbol/symbol.dart';
+
+import '../ads_resolver.dart';
+import 'card.dart';
+import 'placeholder.dart';
+import 'type/color.dart';
+import 'type/number.dart';
+import 'type/gem.dart';
+
+class BigCard extends StatefulWidget {
+  const BigCard({
+    @required this.combination,
+  });
+
+  final PreparedSymbolCombination combination;
+
+  @override
+  BigCardState createState() => BigCardState();
+}
+
+class BigCardState extends State<BigCard> {
+  Map<CardType, Widget> cardType;
+
+  @override
+  void initState() {
+    cardType = {
+      CardType.COLOR: CardTypeColor(
+        first: widget.combination.colors[0],
+        second: widget.combination.colors[1],
+        third: widget.combination.colors[2],
+      ),
+      CardType.NUMBER: CardTypeNumber(
+        numberName: widget.combination.digit,
+      ),
+      CardType.TARROT: CardTypeTarrot(
+        tarrotName: widget.combination.tarrot,
+      ),
+      CardType.GEM: CardTypeGem(
+        gemName: widget.combination.mineral,
+      ),
+      CardType.TEXT: CardTypeProphecy(),
+    };
+
+    if (AppGlobal.ads.adsWatched) CardsLogic.of(context).adsWatched = true;
+
+    /// if need to see ads inside debug mode, comment this line
+    // if (AppGlobal.adsDisabled) CardsLogic.of(context).adsWatched = true;
+
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    if (CardsLogic.of(context).adsWatched) {
+      return PredictionCard(
+        type: cardType[CardsLogic.of(context).choise],
+      );
+    }
+
+    if (!CardsLogic.of(context).adsWatched && CardsLogic.of(context).currentBigCardIsEmpty) {
+      return const CardPlaceholder();
+    }
+
+    if (!CardsLogic.of(context).adsWatched &&
+        !CardsLogic.of(context).currentBigCardIsEmpty &&
+        CardsLogic.of(context).dontShowAds) {
+      return PredictionCard(
+        type: cardType[CardsLogic.of(context).choise],
+      );
+    }
+
+    if (!CardsLogic.of(context).adsWatched && !CardsLogic.of(context).dontShowAds) {
+      return CardsAdsResolver();
+    }
+
+    return PredictionCard(
+      type: cardType[CardsLogic.of(context).choise],
+    );
+  }
+
+  @override
+  void didChangeDependencies() {
+    CardsLogic.of(context).addListener(() {
+      try {
+        setState(() {});
+      } catch (_) {}
+    });
+    super.didChangeDependencies();
+  }
+}
